+++ src/PSWSMan/Commands/PSWSManCertValidationCallback.cs
@@ -6,7 +6,7 @@ using System.Net.Security;
 using System.Collections.ObjectModel;
 using System.Security.Cryptography.X509Certificates;
 
-namespace PSWSMan.Module.Commands;
+namespace PSWSMan.Commands;
 
 [Cmdlet(
     VerbsCommon.New, "PSWSManCertValidationCallback"