+++ src/PSWSMan/Patches/WSManAPIDataCommon.cs
@@ -3,7 +3,7 @@ using System;
 using System.Management.Automation.Remoting.Client;
 using System.Reflection;
 
-namespace PSWSMan.Module.Patches;
+namespace PSWSMan.Patches;
 
 internal static class PSWSMan_WSManApiDataCommon
 {